+++ RoyalAxe/Assets/Scripts/Units/Factory/ISkillFactory.cs
@@ -2,9 +2,9 @@
 {
     public interface ISkillFactory
     {
-        void EquipMobWeapon(UnitsEntity unitsEntity, string weaponId, byte weaponLevel);
+        void EquipMobWeapon(UnitsEntity unitsEntity, string weaponId, int weaponLevel);
 
-        void CreateTestPlayerSkill(UnitsEntity player, string weaponId, byte weaponLevel);
+        void CreateTestPlayerSkill(UnitsEntity player, string weaponId, int weaponLevel);
         void CreateMeleeAttackSkill(UnitsEntity boson, UnitsEntity owner);
     }
 }
